Initial Commit of Classes in Detailed Design
These are the classes outlined in the detailed design. As of now, location and a few other portions of functionality are not complete, specifically the portions that require access to the database.

I will confer with Jack about the creation of the server class, at which point I should be able to implement the portions missing that require server access